latest change

+++ tgaimage.h
@@ -38,7 +38,7 @@ struct TGAColor {
 	TGAColor(unsigned char R, unsigned char G, unsigned char B, unsigned char A) : b(B), g(G), r(R), a(A), bytespp(4) {
 	}
 
-	TGAColor(int v, int bpp) : val(v), bytespp(bpp) {
+	TGAColor(double v, double bpp) : val(v), bytespp(bpp) {
 	}
 
 	TGAColor(const TGAColor &c) : val(c.val), bytespp(c.bytespp) {
